clean up job tabs, moved env settings and meta into configuration tab
