+++ app/components/base/quick-list/no-item.component.ts
@@ -6,7 +6,11 @@ import { Filter } from "app/utils/filter-builder";
 @Component({
     selector: "bl-no-item",
     template: `
-        <div class="no-item-message" *ngIf="status !== loadingStatuses.Loading && itemCount === 0">
+        <div
+            class="no-item-message"
+            [ngClass]="{'slim': _slimline }"
+            *ngIf="status !== loadingStatuses.Loading && itemCount === 0">
+
             <ng-content select="[icon]"></ng-content>
             <ng-content select="[no-filter]" *ngIf="!withFilter"></ng-content>
             <ng-content select="[with-filter]" *ngIf="withFilter"></ng-content>
@@ -26,8 +30,13 @@ export class NoItemComponent {
         this.withFilter = Boolean(filter && !filter.isEmpty());
     }
 
-    public withFilter = false;
+    @Input()
+    public set slimline(slimline: boolean) {
+        this._slimline = slimline;
+    }
 
+    public withFilter = false;
     public loadingStatuses = LoadingStatus;
 
+    private _slimline = false;
 }
