+++ app/components/job/details/index.ts
@@ -1,4 +1,5 @@
+export * from "./job-configuration.component";
+export * from "./job-default.component";
 export * from "./job-details.module";
 export * from "./job-details.component";
-export * from "./job-properties.component";
 export * from "./error-display";
