--- app/components/job/details/job-metadata.html
@@ -1,8 +1,0 @@
-<bl-property-list>
-    <bl-property-group label="Metadata" class="wide">
-        <bl-text-property *ngFor="let metaItem of jobMetadata" label="{{metaItem.name}}" [value]="metaItem.value"></bl-text-property>
-    </bl-property-group>
-</bl-property-list> 
-<bl-no-item [itemCount]="jobMetadata.length">
-    <span no-filter>This job contains no metadata</span>
-</bl-no-item>
